pyproject.toml
[project]
name = "jobai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apify-client>=2.4.1",
    "openai>=2.21.0",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
